backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic==2.11.7
python-multipart==0.0.20
numpy==2.3.2
pillow==11.3.0
rapidocr-onnxruntime==1.4.4
